Initial commit of user created or added assets
Avoids adding files that were generated by Unity or that were generated
by AI.

Files of interest include:

Assets/Scripts/Terrain.cs - This file is responsible for managing the
terrain, generating a quad tree, subdividing and culling tiles, and
performing the instance rendering of the terrain tiles.

Assets/ShadersAndMaterials/TerrainShader.shader - This shader uses
instance rendering to render the terrain tiles. It does a height map
lookup in the vertex shader to set vertex heights.

Assets/Images/TerrainDataSource.txt - This file contains a link to the
site used to generate the source files for height map and albedo. The
files are large and are not included in this repo.